Singleton ammo counter; weapon switching & sounds
Created a singleton to keep ammo count, plus it lets you add and consume
ammos.
Will be useful for the shop, but it's still a WIP (and broken).
Rewritten weapon switching, now the sheep has an array of available
weapons.
Sounds are now played from the shooting weapon (they were played by the
bullets before). So there is only 1 sound source that plays a sound each
shot.